organized the files and added the admin section of the plugin
