+++ public/wp-content/plugins/advThemeMang/advThemeMang.php
@@ -22,61 +22,9 @@ if(file_exists(dirname(__FILE__). '/vendor/autoload.php')){
     require_once dirname(__FILE__) .'/vendor/autoload.php';
 }
 
-use Inc\Activate;
-use Inc\Deactivate;
-use Inc\admin\AdminPages;
-class AdvThemeMangPlugin {
-    public $pluginName;
-    function __construct() {
-        $this->pluginName = plugin_basename(__FILE__);
-        
+define('PLUGIN_PATH', plugin_dir_path(__FILE__));
+define('PLUGIN_URL',  plugin_dir_url(__FILE__));
 
-    }
-    function register() {
-        add_action('admin_enqueue_scripts', [$this,'enque']);
-        add_action('admin_menu', [$this,'admin_pages']);
-        add_filter("plugin_action_links_$this->pluginName", [$this,'settingsLink']);
-    }
-       
-    function activate(){
-        add_action('init', [$this,'customPostType']);
-       Activate::activate();
-
-    }
-    function deactivate(){
-        
-       Deactivate::deactivate();
-
-    
-    }
-
-    function customPostType(){
-        register_post_type('book', ['public'=>true,'label'=>'Books']);
-    }
-    function enque() {
-        
-        wp_enqueue_style('mystyle', plugins_url('/assets/mystyle.css', __FILE__));
-        wp_enqueue_script('myscript', plugins_url('/assets/myscript.js', __FILE__),[]);
-    }
-    function admin_pages() {
-        add_menu_page('Advance Theme Manager', 'advThemeMang', 'manage_options', 'advThemeMang', [$this,'adminIndex'], 'dashicons-store', 110);
-
-    }
-    function adminIndex() {
-        require_once plugin_dir_path(__FILE__) . 'temp/admin.php';
-    }
-    function settingsLink($link) {
-        $settings='<a href="admin.php?page=advThemeMang">Settings</a>';
-        array_push($link,$settings);
-        return $link;
-    }
+if(class_exists('Inc\\Init')){
+    Inc\Init::register_services();
 }
-
-if(class_exists('AdvThemeMangPlugin')){
-
-    $advThemeMangPlugin = new AdvThemeMangPlugin();
-    $advThemeMangPlugin->register();
-}
-
-register_activation_hook(__FILE__, [$advThemeMangPlugin,'activate']);
-register_deactivation_hook(__FILE__, [$advThemeMangPlugin,'deactivate']);
