+++ public/wp-content/plugins/advThemeMang/inc/Base/Activate.php
