+++ public/wp-content/plugins/advThemeMang/inc/Base/Deactivate.php
