--- public/wp-content/plugins/advThemeMang/inc/admin/AdminPages.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Inc\admin;
-
-class AdminPages{
-  
-}
